Fix deploy-operator
The deploy-operator step failed with the latest oc command line from 5.1, because the build script expected the first line of output from "oc login" to be "Login successful." which is no longer the case.  The first line of output is now "WARNING: Using insecure TLS client config. Setting this option is not supported!" and further lines of the output need to be searched to find the "Login successful." line.
Signed-off-by: Tom O'Shea <osheat@us.ibm.com>