Add documentation for creating a VM (#30)
In addition to adding documentation for creating a VM, this PR also fixes a couple issues discovered during validation on the VM. One was a security problem when building the container image. The container user no longer has write access to /opt/ibm/initScripts and so root privilege is required to write files there. A second issue is that the build of the flightservice operator needs to revert to Gradle 7, because its build script uses features that are no longer supported in Gradle 8.

Signed-off-by: Tom O'Shea <osheat@us.ibm.com>